HyperLogLog data structure works well for cardinality less than 10^7 well

+++ src/main/java/cardinality/hash/HashCodeProvider.java
@@ -1,9 +1,7 @@
 package cardinality.hash;
 
-import java.security.NoSuchAlgorithmException;
-
 public interface HashCodeProvider<T> {
 
-    long hashCode(T object);
+    int hashCode(T object);
 
 }
