+++ src/main/java/cardinality/hash/SimpleHashCodeProvider.java
@@ -3,7 +3,7 @@ package cardinality.hash;
 public class SimpleHashCodeProvider<T> implements HashCodeProvider<T> {
 
     @Override
-    public long hashCode(T object) {
+    public int hashCode(T object) {
         return object.hashCode();
     }
